Generated new React app.
Used create-react-app to build a new react app with conventional components (ie. babel) and no configuration. Given the scope of this assignment & my personal inexperience with React specifically, this seemed like a good choice. If I was making a production app I would be more deliberate with the build confiuration. I also emptied out some of the placeholder content that comes with create-react-app.